# feat: optimize span normalization