--- a/pkg/trace/traceutil/normalize/normalize.go
+++ b/pkg/trace/traceutil/normalize/normalize.go
@@ -39,6 +39,20 @@ var (
 	ErrInvalid = errors.New("invalid")
 )
 
+var isAlphaLookup = [256]bool{}
+var isAlphaNumLookup = [256]bool{}
+var isValidASCIIStartCharLookup = [256]bool{}
+var isValidASCIITagCharLookup = [256]bool{}
+
+func init() {
+	for i := 0; i < 256; i++ {
+		isAlphaLookup[i] = isAlpha(byte(i))
+		isAlphaNumLookup[i] = isAlphaNum(byte(i))
+		isValidASCIIStartCharLookup[i] = isValidASCIIStartChar(byte(i))
+		isValidASCIITagCharLookup[i] = isValidASCIITagChar(byte(i))
+	}
+}
+
 // NormalizeName normalizes a span name and returns an error describing the reason
 // (if any) why the name was modified.
 //
@@ -282,7 +296,7 @@ func normMetricNameParse(name string) (string, bool) {
 	res := resa[:0]
 
 	// skip non-alphabetic characters
-	for ; i < len(name) && !isAlpha(name[i]); i++ {
+	for ; i < len(name) && !isAlphaLookup[name[i]]; i++ {
 		continue
 	}
 
@@ -293,7 +307,7 @@ func normMetricNameParse(name string) (string, bool) {
 
 	for ; i < len(name); i++ {
 		switch {
-		case isAlphaNum(name[i]):
+		case isAlphaNumLookup[name[i]]:
 			res = append(res, name[i])
 			ptr++
 		case name[i] == '.':
@@ -334,7 +348,7 @@ func isNormalizedASCIITag(tag string, checkValidStartChar bool) bool {
 	}
 	i := 0
 	if checkValidStartChar {
-		if !isValidASCIIStartChar(tag[0]) {
+		if !isValidASCIIStartCharLookup[tag[0]] {
 			return false
 		}
 		i++
@@ -342,13 +356,13 @@ func isNormalizedASCIITag(tag string, checkValidStartChar bool) bool {
 	for ; i < len(tag); i++ {
 		b := tag[i]
 		// TODO: Attempt to optimize this check using SIMD/vectorization.
-		if isValidASCIITagChar(b) {
+		if isValidASCIITagCharLookup[b] {
 			continue
 		}
 		if b == '_' {
 			// an underscore is only okay if followed by a valid non-underscore character
 			i++
-			if i == len(tag) || !isValidASCIITagChar(tag[i]) {
+			if i == len(tag) || !isValidASCIITagCharLookup[tag[i]] {
 				return false
 			}
 		} else {