### What does this PR do?

These PR does two things:

* Add a non allocating fast path for span name validation. The function `normMetricNameParse` accounts for 3% of all allocation in the trace agent, because it always copies the name, even if it is valid.

This PR also adds benchmarks and tests to check that the fast path really does maintain the correct behaviour

<img width="976" alt="Screenshot 2025-05-06 at 23 55 07" src="https://github.com/user-attachments/assets/8651c9b8-3969-40fd-b398-b9a3b839c773" />

This makes the fast path 4 times as fast, and non allocating

```
BenchmarkNormalizeName

Before
BenchmarkNormalizeName/ok-10 	62699263	        19.03 ns/op	       4 B/op	       1 allocs/op
BenchmarkNormalizeName/truncate-10         	 6011515	       207.3 ns/op	      96 B/op	       1 allocs/op
BenchmarkNormalizeName/remap-10            	51140992	        23.28 ns/op	       8 B/op	       1 allocs/op

After:
BenchmarkNormalizeName/ok-10 	216040237	         5.415 ns/op	       0 B/op	       0 allocs/op
BenchmarkNormalizeName/truncate-10         	 8589586	       138.4 ns/op	      96 B/op	       1 allocs/op
BenchmarkNormalizeName/remap-10            	51851065	        23.23 ns/op	       8 B/op	       1 allocs/op
```

* Use a lookup table to check ascii character ranges

This is slightly (15%) faster in the fast path

```
BenchNormalizeTag

Before
BenchmarkNormalizeTag/ok-10 	161304495	         7.763 ns/op	       0 B/op	       0 allocs/op
BenchmarkNormalizeTag/trim-10         	14470398	        80.40 ns/op	      32 B/op	       2 allocs/op
BenchmarkNormalizeTag/trim-both-10    	 9167113	       131.8 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTag/plenty-10       	 9269030	       136.4 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTag/more-10         	 5955565	       205.7 ns/op	     128 B/op	       4 allocs/op

After
BenchmarkNormalizeTag/ok-10 	176645220	         6.654 ns/op	       0 B/op	       0 allocs/op
BenchmarkNormalizeTag/trim-10         	15125716	        79.96 ns/op	      32 B/op	       2 allocs/op
BenchmarkNormalizeTag/trim-both-10    	 9216984	       130.1 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTag/plenty-10       	 9756668	       124.8 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTag/more-10         	 5990361	       198.6 ns/op	     128 B/op	       4 allocs/op

```

```
BenchmarkNormalizeTagValue

Before
BenchmarkNormalizeTagValue/ok-10 	154064102	         7.926 ns/op	       0 B/op	       0 allocs/op
BenchmarkNormalizeTagValue/trim-10         	11178790	       108.3 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTagValue/trim-both-10    	 7605100	       161.9 ns/op	     128 B/op	       4 allocs/op
BenchmarkNormalizeTagValue/plenty-10       	 9153663	       126.8 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTagValue/more-10         	 6018870	       199.6 ns/op	     128 B/op	       4 allocs/op
BenchmarkNormalizeTagValue/digit-10        	122968783	         9.862 ns/op	       0 B/op	       0 allocs/op

After
BenchmarkNormalizeTagValue/ok-10 	170819376	         6.933 ns/op	       0 B/op	       0 allocs/op
BenchmarkNormalizeTagValue/trim-10         	11231467	       106.7 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTagValue/trim-both-10    	 7575480	       159.8 ns/op	     128 B/op	       4 allocs/op
BenchmarkNormalizeTagValue/plenty-10       	 9601680	       126.0 ns/op	      64 B/op	       3 allocs/op
BenchmarkNormalizeTagValue/more-10         	 6015847	       199.3 ns/op	     128 B/op	       4 allocs/op
BenchmarkNormalizeTagValue/digit-10        	100000000	        10.13 ns/op	       0 B/op	       0 allocs/op
```

### Motivation

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes. If you validated changes manually and not through automated tests, add context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask reviewers to do it. Describe how to set up an environment for manual tests in the PR description. Manual validation is expected to happen on every commit before merge.

Any manual validation step should then map to an automated test. Manual validation should not substitute automation, minus exceptions not supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->